requirements.txt
ccy>=0.3.2
python-stdnet
dynts
unuk >= 0.2.2
django-piston